+++ data-raw/stations.R
@@ -5,7 +5,7 @@ require(readr)
 require(lubridate)
 
 stations <- read_excel(file.path("data-raw", "stations.xlsx"), sheet="lat_long")%>%
-  mutate(Source=recode(Source, TNS="STN"))
+  mutate(Source=recode(Source, TNS="STN", twentymm="20mm"))
 
 stationsEMPEZ<-read_csv(file.path("data-raw", "EZ_stations.csv"),
                         col_types = cols_only(SampleDate="c", StationCode="c", Lat="d", Long="d"))%>%