+++ tests/testthat/test-Data.R
@@ -4,7 +4,7 @@ require(stringr)
 require(lubridate)
 
 Data_sets <- c("EMP_Meso", "FMWT_Meso", "STN_Meso",
-               "twentymm_Meso", "FRP_Meso","EMP_Micro",
+               "20mm_Meso", "FRP_Meso","EMP_Micro",
                "FRP_Macro", "EMP_Macro", "FMWT_Macro", "STN_Macro"
                )
 
@@ -32,7 +32,7 @@ test_that("Not all CPUEs are 0", {
 })
 
 test_that("Only the expected stations are missing coordinates", {
-  expect_setequal(unique(No_coords$Station), c("twentymm 798", "twentymm 799", "twentymm 794", "twentymm 795", "twentymm 796", "twentymm 797"))
+  expect_setequal(unique(No_coords$Station), c("20mm 798", "20mm 799", "20mm 794", "20mm 795", "20mm 796", "20mm 797"))
 })
 
 test_that("Date and Datetime and displaying the same dates", {
@@ -42,4 +42,8 @@ test_that("Date and Datetime and displaying the same dates", {
 test_that("Bottom depths are within reasonable limits", {
   expect_true(all(zoopEnvComb$BottomDepth > 0.5 | is.na(zoopEnvComb$BottomDepth)))
   expect_true(all(zoopEnvComb$BottomDepth < 35 | is.na(zoopEnvComb$BottomDepth)))
+})
+
+test_that("There are no NA Volumes in zoopComb", {
+  expect_equal(length(which(is.na(zoopComb$Volume))), 0)
 })