+++ tests/testthat/test-Zoopdownloader.R
@@ -4,8 +4,6 @@ Data_sets <- c("EMP_Meso", "FMWT_Meso", "STN_Meso",
               "20mm_Meso", "FRP_Meso","EMP_Micro",
               "FRP_Macro", "EMP_Macro", "FMWT_Macro", "STN_Macro"
               )
-Data_sets2 <- Data_sets
-Data_sets2[which(Data_sets2=="20mm_Meso")] <- "twentymm_Meso"
 
 Data <- Zoopdownloader(Data_folder = tempdir(), Return_object = TRUE,
                          Save_object = FALSE, Redownload_data = TRUE,
@@ -15,7 +13,7 @@ No_coords2<-dplyr::filter(Data$Environment, is.na(Latitude) & !stringr::str_dete
   dplyr::mutate(Station=paste(Source, Station))
 
 test_that("Dowloaded data includes all datasets", {
-  expect_setequal(unique(paste(Data$Zooplankton$Source, Data$Zooplankton$SizeClass, sep="_")), Data_sets2)
+  expect_setequal(unique(paste(Data$Zooplankton$Source, Data$Zooplankton$SizeClass, sep="_")), Data_sets)
 })
 
 test_that("No samples duplicated", {
@@ -31,7 +29,7 @@ test_that("Not all CPUEs are 0", {
 })
 
 test_that("Only the expected stations are missing coordinates", {
-  expect_setequal(unique(No_coords2$Station), c("twentymm 798", "twentymm 799", "twentymm 794", "twentymm 795", "twentymm 796", "twentymm 797"))
+  expect_setequal(unique(No_coords2$Station), c("20mm 798", "20mm 799", "20mm 794", "20mm 795", "20mm 796", "20mm 797"))
 })
 
 test_that("Date and Datetime and displaying the same dates", {
@@ -41,4 +39,8 @@ test_that("Date and Datetime and displaying the same dates", {
 test_that("Bottom depths are within reasonable limits", {
   expect_true(all(Data$Environment$BottomDepth > 0.5 | is.na(Data$Environment$BottomDepth)))
   expect_true(all(Data$Environment$BottomDepth < 35 | is.na(Data$Environment$BottomDepth)))
+})
+
+test_that("There are no NA Volumes in Zoopdownloader output", {
+  expect_equal(length(which(is.na(Data$Zooplankton$Volume))), 0)
 })